Move the PrependGroupingObject function into the prune package
